feat: Refactor timeout handling into TimeoutClient component and simplify TimeoutPage
